+++ src/Application/Async/Http.php
@@ -49,6 +49,7 @@ final class Http implements Implementation
      * @param \Closure(OperatingSystem, Environment): Builder $container
      * @param Sequence<callable(Pipe, Container): Component<SideEffect, Response>> $routes
      * @param \Closure(Component<SideEffect, Response>, Container): Component<SideEffect, Response> $mapRoute
+     * @param \Closure(Component<SideEffect, Response>, Container): Component<SideEffect, Response> $mapRoutes
      * @param Maybe<callable(ServerRequest, Container): Attempt<Response>> $notFound
      * @param \Closure(ServerRequest, \Throwable, Container): Attempt<Response> $recover
      */
@@ -58,6 +59,7 @@ final class Http implements Implementation
         private \Closure $container,
         private Sequence $routes,
         private \Closure $mapRoute,
+        private \Closure $mapRoutes,
         private Maybe $notFound,
         private \Closure $recover,
     ) {
@@ -76,6 +78,7 @@ final class Http implements Implementation
             static fn(OperatingSystem $os, Environment $env) => [$os, $env],
             static fn() => Builder::new(),
             Sequence::lazyStartingWith(),
+            static fn(Component $component) => $component,
             static fn(Component $component) => $component,
             $notFound,
             static fn(ServerRequest $request, \Throwable $e) => Attempt::error($e),
@@ -101,6 +104,7 @@ final class Http implements Implementation
             $this->container,
             $this->routes,
             $this->mapRoute,
+            $this->mapRoutes,
             $this->notFound,
             $this->recover,
         );
@@ -125,6 +129,7 @@ final class Http implements Implementation
             $this->container,
             $this->routes,
             $this->mapRoute,
+            $this->mapRoutes,
             $this->notFound,
             $this->recover,
         );
@@ -147,6 +152,7 @@ final class Http implements Implementation
             ),
             $this->routes,
             $this->mapRoute,
+            $this->mapRoutes,
             $this->notFound,
             $this->recover,
         );
@@ -182,6 +188,7 @@ final class Http implements Implementation
             $this->container,
             ($this->routes)($handle),
             $this->mapRoute,
+            $this->mapRoutes,
             $this->notFound,
             $this->recover,
         );
@@ -204,6 +211,30 @@ final class Http implements Implementation
                 $previous($component, $get),
                 $get,
             ),
+            $this->mapRoutes,
+            $this->notFound,
+            $this->recover,
+        );
+    }
+
+    /**
+     * @psalm-mutation-free
+     */
+    #[\Override]
+    public function mapRoutes(callable $map): self
+    {
+        $previous = $this->mapRoutes;
+
+        return new self(
+            $this->os,
+            $this->map,
+            $this->container,
+            $this->routes,
+            $this->mapRoute,
+            static fn($component, $get) => $map(
+                $previous($component, $get),
+                $get,
+            ),
             $this->notFound,
             $this->recover,
         );
@@ -221,6 +252,7 @@ final class Http implements Implementation
             $this->container,
             $this->routes,
             $this->mapRoute,
+            $this->mapRoutes,
             Maybe::just($handle),
             $this->recover,
         );
@@ -240,6 +272,7 @@ final class Http implements Implementation
             $this->container,
             $this->routes,
             $this->mapRoute,
+            $this->mapRoutes,
             $this->notFound,
             static fn($request, $e, $container) => $previous($request, $e, $container)->recover(
                 static fn($e) => $recover($request, $e, $container),
@@ -255,6 +288,7 @@ final class Http implements Implementation
         $routes = $this->routes;
         $notFound = $this->notFound;
         $mapRoute = $this->mapRoute;
+        $mapRoutes = $this->mapRoutes;
         $recover = $this->recover;
 
         $run = Commands::of(Serve::of(
@@ -265,6 +299,7 @@ final class Http implements Implementation
                 $routes,
                 $notFound,
                 $mapRoute,
+                $mapRoutes,
                 $recover,
             ): Response {
                 $env = Environment::of($env);
@@ -275,6 +310,7 @@ final class Http implements Implementation
                     ->map(static fn($handle) => $handle($pipe, $container))
                     ->map(static fn($component) => $mapRoute($component, $container));
                 $router = new Router(
+                    static fn($component) => $mapRoutes($component, $container),
                     $routes,
                     $notFound->map(
                         static fn($handle) => static fn(ServerRequest $request) => $handle(